+++ src/ui/tabs/OtherTab.ts
@@ -1,12 +1,8 @@
 import { ObsidianLayout, add, toast, CENTER } from "frida-java-menu";
 
 import { I18n } from "../../i18n/I18n";
-import { State } from "../../data/State";
-import { ModuleManager } from "../../core/ModuleManager";
 import { JavaUtils } from "../../utils/JavaUtils";
-import { UnityUtils } from "../../utils/UnityUtils";
 import { Constants } from "../../data/Constants";
-import { UwUifyModule } from "../../modules/UwUify";
 
 export class OtherTab {
     static draw(layout: ObsidianLayout) {
@@ -26,15 +22,5 @@ export class OtherTab {
         add(layout.button(I18n.t("menu.other.discord_url"), () => JavaUtils.openURL(Constants.DISCORD_URL)));
         add(layout.button(I18n.t("menu.other.malummenu_url"), () => JavaUtils.openURL(Constants.MALUMMENU_URL)));
         add(layout.button(I18n.t("menu.other.changelog"), () => JavaUtils.openURL(Constants.GITHUB_CHANGELOG_URL)));
-
-        add(
-            layout.toggle(
-                I18n.t("menu.functions.uwuify"),
-                UnityUtils.run((state: boolean) => {
-                    State.uwuifyMode = state;
-                    ModuleManager.get(UwUifyModule)?.toggleUwUify(state);
-                })
-            )
-        );
     }
 }