+++ docs/CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/2.0.0/),
 
 ### Changed
 
+- Move UwUify Game from Other Tab to Passive Tab
 - Bump frida from 17.16.3 to 17.16.4 by [@dependabot] ([#11])
 - Bump frida from 17.15.5 to 17.16.3 by [@dependabot] ([#10])
 - Bump frida from 17.15.4 to 17.15.5 by [@dependabot] ([#7])